--- web_hackaton/src/main/java/ru/atom/chat/User.java
@@ -1,17 +1,0 @@
-package ru.atom.chat;
-
-public class User {
-    private String name;
-
-    public User(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
