+++ web_hackaton/src/main/java/ru/atom/chat/models/User.java
@@ -1,0 +1,31 @@
+package ru.atom.chat.models;
+
+public class User {
+    private String name;
+    private boolean isOnline;
+
+    public boolean isOnline() {
+        return isOnline;
+    }
+
+    public void setOnline(boolean online) {
+        isOnline = online;
+    }
+
+    public User(String name, boolean isOnline) {
+        this.name = name;
+        this.isOnline = isOnline;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public User(String name) {
+        this.name = name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
